picam integration next

+++ mcp.h
@@ -108,23 +108,23 @@ typedef struct mcp_sensors_t {
 	uint16_t bh1750_lux_mean;
 
 	// BMP085 calibration data
-	int16_t bmp085_ac1;
-	int16_t bmp085_ac2;
-	int16_t bmp085_ac3;
-	uint16_t bmp085_ac4;
-	uint16_t bmp085_ac5;
-	uint16_t bmp085_ac6;
-	int16_t bmp085_b1;
-	int16_t bmp085_b2;
-	int16_t bmp085_mb;
-	int16_t bmp085_mc;
-	int16_t bmp085_md;
+	int bmp085_ac1;
+	int bmp085_ac2;
+	int bmp085_ac3;
+	unsigned int bmp085_ac4;
+	unsigned int bmp085_ac5;
+	unsigned int bmp085_ac6;
+	int bmp085_b1;
+	int bmp085_b2;
+	int bmp085_mb;
+	int bmp085_mc;
+	int bmp085_md;
 
 	// BMP085 temperature + barometric pressure
 	float bmp085_temp;
-	uint16_t bmp085_utemp;
+	unsigned int bmp085_utemp;
 	float bmp085_baro;
-	uint16_t bmp085_ubaro;
+	unsigned int bmp085_ubaro;
 
 	// BMP280 temperature + barometric pressure
 	float bmp280_temp;
